shoes-bot-backend/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
requests
aiogram